+++ activesupport/test/broadcast_logger_test.rb
@@ -114,14 +114,6 @@ module ActiveSupport
       assert_equal [[::Logger::FATAL, "seen", nil]], log2.adds
     end
 
-    test "Including top constant LoggerSilence is deprecated" do
-      assert_deprecated("Please use `ActiveSupport::LoggerSilence`") do
-        Class.new(CustomLogger) do
-          include ::LoggerSilence
-        end
-      end
-    end
-
     class CustomLogger
       include ActiveSupport::LoggerSilence